requirements.txt
mcp>=1.0.0
pydantic>=2.0.0
pydantic-ai>=0.4.0
asyncio-mqtt>=0.16.0
httpx>=0.25.0
python-dotenv>=1.0.0
firecrawl-py>=0.2.7
openai>=1.30.1
tiktoken>=0.6.0
supabase
psycopg2-binary>=2.9.9
markdownify